feat(Utils): make recursive delete

+++ APIServer/src/Utils.ts
@@ -1,5 +1,6 @@
 import * as fs from "fs";
 import express = require("express");
+import { spawn } from "child_process";
 
 export function isReadableFile(absPath: string): Promise<boolean> {
     return new Promise((resolve) => {
@@ -9,16 +10,53 @@ export function isReadableFile(absPath: string): Promise<boolean> {
     });
 }
 
-export function deleteDir(absPath: string): Promise<boolean> {
-    return new Promise((resolve, reject) => {
-        fs.rmdir(absPath, (err) => {
-            if (err) {
-                reject(err);
-            } else {
-                resolve(false);
-            }
+export function deleteDir(
+    absPath: string,
+    recursive: boolean = false,
+): Promise<boolean> {
+    if (recursive) {
+        return new Promise((resolve, reject) => {
+            const args = [
+                "-r", // treat ALL files as text
+                "--", // never overwrite existing files
+                absPath, // path
+            ];
+            console.log("rm", args);
+            const child = spawn("rm", args, {
+                env: { LC_ALL: "C" },
+                stdio: ["pipe", "pipe", "pipe"],
+            });
+
+            child.stdout.on("data", (data) => {
+                console.log(`stdout: ${data}`);
+            });
+
+            child.stderr.on("data", (data) => {
+                console.log(`err: ${data}`);
+            });
+
+            child.on("error", reject);
+
+            child.on("close", (code) => {
+                console.log(`unzip exited with code ${code}`);
+                if (code === 0) {
+                    resolve(true);
+                } else {
+                    reject(false);
+                }
+            });
         });
-    });
+    } else {
+        return new Promise((resolve, reject) => {
+            fs.rmdir(absPath, (err) => {
+                if (err) {
+                    reject(err);
+                } else {
+                    resolve(false);
+                }
+            });
+        });
+    }
 }
 
 export function asyncMiddleware(fn: express.RequestHandler) {
